# server/documentation/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cosmian Authentication Verifier
  version: 1.0.0
  description: |
    High-performance authentication and session management server.
    All endpoints are served over **HTTPS only**.

    ## Authentication for admin endpoints
    Admin endpoints require a session cookie obtained by logging into the `_` realm:
    ```
    POST /login?realm=_
    Authorization: Basic <base64(username:password)>
    ```
    The server responds with `Set-Cookie: _ea_=<value>; HttpOnly; Secure; SameSite=Strict`.
    Pass this cookie on every subsequent admin request.
servers:
  - url: https://localhost:8443
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: _ea_
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    mtls:
      type: mutualTLS
    appToken:
      type: apiKey
      in: header
      name: X-Vault-Token
      description: |
        AppRole-compatible bearer token issued by this server's AppRole or
        Kubernetes login endpoints.  Required on `/auth/token/*` endpoints.
  schemas:
...
    TokenResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: RFC 9068 `at+jwt` access token.
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Access-token lifetime in seconds.
        scope:
          type: string
          description: Space-delimited granted scope.
        id_token:
          type: string
          description: OIDC ID token (present when the `openid` scope was granted).
        refresh_token:
          type: string
          description: Opaque refresh token (present when offline access is granted).
    OAuthError:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: invalid_grant
        error_description:
          type: string
...
paths:
...
  /oidc/token:
    post:
      tags: [OIDC]
      summary: Token endpoint
      operationId: oidcToken
      description: |
        Issues tokens for the `authorization_code` (with mandatory PKCE),
        `refresh_token` (rotating, with reuse detection), and
        `client_credentials` grants. Client authentication via
        `client_secret_basic` or `client_secret_post` (public clients use PKCE).
      security:
        - basicAuth: []
        - {}
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type]
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code, refresh_token, client_credentials]
                code: {type: string}
                redirect_uri: {type: string}
                code_verifier: {type: string}
                refresh_token: {type: string}
                scope: {type: string}
                client_id: {type: string}
                client_secret: {type: string}
      responses:
        '200':
          description: Token response.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: OAuth error (invalid_request, invalid_grant, ...).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
        '401':
          description: Invalid client authentication.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OAuthError'
  /oidc/userinfo:
    get:
      tags: [OIDC]
      summary: UserInfo endpoint
      operationId: oidcUserinfoGet
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Scope-gated claims for the token subject.
          content:
            application/json:
              schema: {type: object}
        '401':
          description: Missing or invalid access token.
    post:
      tags: [OIDC]
      summary: UserInfo endpoint (POST)
      operationId: oidcUserinfoPost
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Scope-gated claims for the token subject.
          content:
            application/json:
              schema: {type: object}
        '401':
          description: Missing or invalid access token.
  /oidc/introspect:
    post:
      tags: [OIDC]
      summary: Token introspection (RFC 7662)
      operationId: oidcIntrospect
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token: {type: string}
                token_type_hint: {type: string}
      responses:
        '200':
          description: Introspection response (`active` plus metadata when active).
          content:
            application/json:
              schema: {type: object}
  /oidc/revoke:
    post:
      tags: [OIDC]
      summary: Token revocation (RFC 7009)
      operationId: oidcRevoke
      security:
        - basicAuth: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [token]
              properties:
                token: {type: string}
                token_type_hint: {type: string}
      responses:
        '200':
          description: Always 200 OK, regardless of whether the token existed.
...